Add files via upload
The article in the ABC news dated the 21 March 2021, discusses increases in the number family violence incidents in Victoria during the 2020 COVID 19 pandemic. After seeing the article, we were motivated to do our own investigation and conduct some statistical analysis of our own, to see if what was reported in the article had statistical merit.